Site/scripts/requirements.txt
fire==0.4.0
requests==2.28.1
pyyaml==6.0
chardet==5.0.0
python-slugify==6.1.2